requirements.txt
fastapi
langchain
langchain_cohere
streamlit
langserve[all]
python-dotenv